Move ClassDefBuilder to its own file
Otherwise it's sitting bizarrely in the midst of lots of unrelated code.
