+++ Lib/fontTools/otlLib/classDefBuilder.py
@@ -1,0 +1,60 @@
+from fontTools.ttLib.tables import otTables as ot
+
+
+class ClassDefBuilder(object):
+    """Helper for building ClassDef tables."""
+    def __init__(self, useClass0):
+        self.classes_ = set()
+        self.glyphs_ = {}
+        self.useClass0_ = useClass0
+
+    def canAdd(self, glyphs):
+        if isinstance(glyphs, (set, frozenset)):
+            glyphs = sorted(glyphs)
+        glyphs = tuple(glyphs)
+        if glyphs in self.classes_:
+            return True
+        for glyph in glyphs:
+            if glyph in self.glyphs_:
+                return False
+        return True
+
+    def add(self, glyphs):
+        if isinstance(glyphs, (set, frozenset)):
+            glyphs = sorted(glyphs)
+        glyphs = tuple(glyphs)
+        if glyphs in self.classes_:
+            return
+        self.classes_.add(glyphs)
+        for glyph in glyphs:
+            assert glyph not in self.glyphs_
+            self.glyphs_[glyph] = glyphs
+
+    def classes(self):
+        # In ClassDef1 tables, class id #0 does not need to be encoded
+        # because zero is the default. Therefore, we use id #0 for the
+        # glyph class that has the largest number of members. However,
+        # in other tables than ClassDef1, 0 means "every other glyph"
+        # so we should not use that ID for any real glyph classes;
+        # we implement this by inserting an empty set at position 0.
+        #
+        # TODO: Instead of counting the number of glyphs in each class,
+        # we should determine the encoded size. If the glyphs in a large
+        # class form a contiguous range, the encoding is actually quite
+        # compact, whereas a non-contiguous set might need a lot of bytes
+        # in the output file. We don't get this right with the key below.
+        result = sorted(self.classes_, key=lambda s: (len(s), s), reverse=True)
+        if not self.useClass0_:
+            result.insert(0, frozenset())
+        return result
+
+    def build(self):
+        glyphClasses = {}
+        for classID, glyphs in enumerate(self.classes()):
+            if classID == 0:
+                continue
+            for glyph in glyphs:
+                glyphClasses[glyph] = classID
+        classDef = ot.ClassDef()
+        classDef.classDefs = glyphClasses
+        return classDef
